% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams.R
\docType{data}
\name{cams}
\alias{cams}
\title{Dati cams gennaio-maggio 2020}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 124944 rows and 8 columns.
}
\usage{
cams
}
\description{
Dati di no2, pm10 e pm25, serie counterfactual di CAMS
}
\keyword{datasets}
