% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricaDati.R
\name{caricaDati}
\alias{caricaDati}
\title{Carica i dati di un inquinante}
\usage{
caricaDati(pacchetto, inquinante, verbose = FALSE)
}
\arguments{
\item{pacchetto}{Nome del pacchetto, stringa di caratteri}

\item{inquinante}{Nome dell'inquinante, stringa di caratteri}

\item{verbose}{Logical, info aggiuntive}
}
\value{
tibble con i dati richiesti
}
\description{
Questa funzione permette di caricare i dati di un determinato inquinante in un determinato pacchetto (regione), senza dover caricare il pacchetto. La funzione richiede
come parametri il nome della regione e dell'inquinante. Questa funzione e' pensata per facilitare il riutilizzo di uno stesso script su piu' regioni e su piu' inquinanti
}
