% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_km.R
\name{in_km}
\alias{in_km}
\title{Trasforma le coordinate da metri a km}
\usage{
in_km(.x)
}
\arguments{
\item{.x}{oggetto sf}
}
\value{
oggetto sf con coordinate in kilometri
}
\description{
Per dati spaziali con epsg=32632 trasforma lecoordinate da metri a kilometri
}
