% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installa_pacchetti.R
\name{installa_pacchetto_pulvirus}
\alias{installa_pacchetto_pulvirus}
\title{Funzione per installare i pacchetti degli inquinante}
\usage{
installa_pacchetto_pulvirus(.nomePacchetto, ...)
}
\arguments{
\item{.nomePacchetto}{nome del pacche da installare}

\item{...}{Altri parametri per la funzione install_github}
}
\description{
La funzione installa il pacchetto dal repository github progettopulvirus
}
