% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numero_giorni.R
\name{numero_giorni}
\alias{numero_giorni}
\title{Numero di giorni nel periodo di analisi}
\usage{
numero_giorni(.x)
}
\arguments{
\item{.x}{Un tibble con campo date che descrive la data delle osservazioni}
}
\value{
Il numero di giorni del periodo di analisi
}
\description{
Il numero di giorni tra la data di inizio e di fine nel campo date del tibble di input
}
