% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulvirus_regioni.R
\docType{data}
\name{PULVIRUS_REGIONI}
\alias{PULVIRUS_REGIONI}
\title{Pacchetti dati progetto PULVIRUS}
\format{
An object of class \code{character} of length 18.
}
\usage{
PULVIRUS_REGIONI
}
\description{
Vettore di stringhe con i nomi dei pacchetti che contengono i dati regionali degli inquinanti preparati per il progetto PULVIRUS
}
\keyword{datasets}
