% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_dati.R
\name{prepara_dati}
\alias{prepara_dati}
\title{Funzione per preparare i dati all'analisi del modello}
\usage{
prepara_dati(
  .x,
  previous = FALSE,
  logaritmo = TRUE,
  lockdown = TRUE,
  inizio_lockdown = "2020-03-09",
  fine_lockdown = "2020-05-03",
  day = FALSE,
  wday = FALSE,
  week = FALSE,
  weekend = FALSE
)
}
\arguments{
\item{.x}{Un tibble con i dati di input}

\item{previous}{Logical}

\item{logaritmo}{Logical}

\item{lockdown}{Logical}

\item{inizio_lockdown}{Character o Logical nel formato \%Y-\%m-\%d}

\item{fine_lockdown}{Character o Logical nel formato \%Y-\%m-\%d}

\item{day}{Logical}

\item{wday}{Logical}

\item{week}{Logical}

\item{weekend}{Logical}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
La funzione restituisce il tibble .x con in piu': 1) la variabile "banda", un intero che va da 1 al numero di osservazioni per il periodo
oggetto di analisi; la variabile Intercept. La variabile "banda" e' necessaria per costruire il campo SPDE.
Settando il parametro previous a TRUE viene restituito inoutput un tibble con la colonna "pvalue" ovvero il valore dell'inquinante nel giorno precedente.
Mettendo il parametro logaritmo=TRUE il tibble .x presenta in output la variabile lvalue (logaritmo di value). Se previous e' uguale a TRUE viene restituita anche la colonna
lpvalue ovvero il logaritmo del valore dell'inquinante nel giorno precedente.
Il parametro lockdown=TRUE restitusce nel tibble di output una colonna di nome "lockdown" pari a 1 per il periodo compreso tra "inizio_lockdown" e "fine_lockdown"
pari a 0 altrimenti. Impostando day a TRUE viene creato un duplicato della variabile banda (banda e' una variabile pensata per l'spde mentre
day puo' essere usata per uno smoother al di fuori dell'spde). Impostando wday a TRUE viene creata una variabile che assume valori da 1 a 7 a seconda del giorno della settimana.
Impostando week a TRUE viene creata una variabile che assume valori da 1 a numero delle settimane del periodo in esame.
Impostando weekend a TRUE viene creata una variabile 0/1 per l'effetto fine settimana (sabato-domenica).
}
